feat: :sparkles: add sign in screen
