+++ src/components/AuthWidget/SignUpWidget/ui/SuccessScreen/index.tsx
